+++ backend-core/src/main/java/com/promptpicture/backend/core/cart/adapter/CartRepositoryAdapter.java
@@ -1,13 +1,14 @@
 package com.promptpicture.backend.core.cart.adapter;
 
 import com.promptpicture.backend.core.cart.domain.Cart;
+import com.promptpicture.backend.core.cart.domain.CustomerCartItem;
 
 import java.util.List;
 import java.util.UUID;
 
 public interface CartRepositoryAdapter {
 
-    void addPromptToCart(UUID externalCustomerId, Long promptId);
+    void addPromptToCart(CustomerCartItem customerCartItem);
 
     Cart getCartByExternalCustomerId(UUID externalCustomerId);