+++ entrypoints/prompt-rest-api/src/main/java/com/promptpicture/backend/entrypoint/rest/model/input/CartItemRequest.java
@@ -9,5 +9,6 @@ public class CartItemRequest {
 
     private UUID externalCustomerId;
     private Long promptId;
+    private String resolution;
 
 }