+++ output-adapters/jpa-db-entities/src/main/java/com/promptpicture/backend/jpa/prompt/repository/PromptDataEntityRepository.java
@@ -1,0 +1,8 @@
+package com.promptpicture.backend.jpa.prompt.repository;
+
+import com.promptpicture.backend.jpa.prompt.entity.PromptOrderDataEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PromptDataEntityRepository extends JpaRepository<PromptOrderDataEntity, Long> {
+
+}